app/requirements.txt
 django 
 django-environ 
 djangorestframework 
 django-cors-headers 
 djangorestframework-simplejwt

 isort
 autoflake